+++ client/branded/src/search-ui/components/OwnerSearchResult.tsx
@@ -4,6 +4,9 @@ import classNames from 'classnames'
 
 import { TeamAvatar } from '@sourcegraph/shared/src/components/TeamAvatar'
 import { UserAvatar } from '@sourcegraph/shared/src/components/UserAvatar'
+import { BuildSearchQueryURLParameters, QueryState, SearchContextProps } from '@sourcegraph/shared/src/search'
+import { FilterKind, findFilter } from '@sourcegraph/shared/src/search/query/query'
+import { appendFilter, omitFilter } from '@sourcegraph/shared/src/search/query/transformer'
 import { getOwnerMatchUrl, OwnerMatch } from '@sourcegraph/shared/src/search/stream'
 import { TelemetryProps } from '@sourcegraph/shared/src/telemetry/telemetryService'
 import { Link } from '@sourcegraph/wildcard'
@@ -13,20 +16,27 @@ import { ResultContainer } from './ResultContainer'
 import styles from './OwnerSearchResult.module.scss'
 import resultStyles from './SearchResult.module.scss'
 
-export interface PersonSearchResultProps extends TelemetryProps {
+export interface OwnerSearchResultProps extends Pick<SearchContextProps, 'selectedSearchContextSpec'>, TelemetryProps {
     result: OwnerMatch
     onSelect: () => void
     containerClassName?: string
     as?: React.ElementType
     index: number
+
+    // If not provided, the result will not contain a link to the owner's files.
+    queryState?: QueryState
+    buildSearchURLQueryFromQueryState?: (queryParameters: BuildSearchQueryURLParameters) => string
 }
 
-export const OwnerSearchResult: React.FunctionComponent<PersonSearchResultProps> = ({
+export const OwnerSearchResult: React.FunctionComponent<OwnerSearchResultProps> = ({
     result,
     onSelect,
     containerClassName,
     as,
     index,
+    queryState,
+    buildSearchURLQueryFromQueryState,
+    selectedSearchContextSpec,
     telemetryService,
 }) => {
     const displayName = useMemo(() => {
@@ -49,6 +59,30 @@ export const OwnerSearchResult: React.FunctionComponent<PersonSearchResultProps>
         }
         return url
     }, [result])
+
+    const fileSearchLink = useMemo(() => {
+        if (!queryState || !buildSearchURLQueryFromQueryState) {
+            return ''
+        }
+
+        const handle = result.handle || result.email
+        if (!handle) {
+            return ''
+        }
+
+        let query = queryState.query
+        const selectFilter = findFilter(queryState.query, 'select', FilterKind.Global)
+        if (selectFilter && selectFilter.value?.value === 'file.owners') {
+            query = omitFilter(query, selectFilter)
+        }
+        query = appendFilter(query, 'file', `has.owner(${handle})`)
+
+        const searchParams = buildSearchURLQueryFromQueryState({
+            query,
+            searchContextSpec: selectedSearchContextSpec,
+        })
+        return `/search?${searchParams}`
+    }, [buildSearchURLQueryFromQueryState, queryState, result.email, result.handle, selectedSearchContextSpec])
 
     const logSearchOwnerClicked = (): void => {
         if (url.startsWith('mailto:')) {
@@ -106,9 +140,14 @@ export const OwnerSearchResult: React.FunctionComponent<PersonSearchResultProps>
                 className={classNames(resultStyles.searchResultMatch, 'p-2 flex-column')}
                 data-testid="owner-search-result"
             >
-                <div className={resultStyles.matchType}>
-                    <small>Owner match</small>
-                </div>
+                <small className={resultStyles.matchType}>
+                    <span>Owner match</span>
+                    {fileSearchLink && (
+                        <Link to={fileSearchLink} className={styles.filesLink}>
+                            Show files
+                        </Link>
+                    )}
+                </small>
                 {result.type === 'person' && !result.user && (
                     <>
                         <div className={resultStyles.dividerVertical} />