+++ client/web/src/enterprise/repo/settings/backend.ts
@@ -1,47 +1,45 @@
-import { Observable } from 'rxjs'
-import { map } from 'rxjs/operators'
+import { gql } from '@sourcegraph/http-client'
 
-import { gql, dataOrThrowErrors } from '@sourcegraph/http-client'
-import { Scalars } from '@sourcegraph/shared/src/graphql-operations'
+export const RepoPermissionsInfoQuery = gql`
+    query RepoPermissionsInfo($repoID: ID!, $first: Int, $last: Int, $after: String, $before: String, $query: String) {
+        node(id: $repoID) {
+            __typename
+            ... on Repository {
+                ...RepoPermissionsInfoRepoNode
+            }
+        }
+    }
 
-import { requestGraphQL } from '../../../backend/graphql'
-import {
-    RepoPermissionsInfoFields,
-    RepoPermissionsInfoResult,
-    RepoPermissionsInfoVariables,
-} from '../../../graphql-operations'
-
-export function repoPermissionsInfo(repo: Scalars['ID']): Observable<RepoPermissionsInfoFields | null> {
-    return requestGraphQL<RepoPermissionsInfoResult, RepoPermissionsInfoVariables>(
-        gql`
-            query RepoPermissionsInfo($repo: ID!) {
-                node(id: $repo) {
-                    __typename
-                    ... on Repository {
-                        permissionsInfo {
-                            ...RepoPermissionsInfoFields
-                        }
-                    }
+    fragment RepoPermissionsInfoRepoNode on Repository {
+        permissionsInfo {
+            syncedAt
+            updatedAt
+            unrestricted
+            users(first: $first, last: $last, after: $after, before: $before, query: $query) {
+                nodes {
+                    ...PermissionsInfoUserFields
+                }
+                totalCount
+                pageInfo {
+                    hasNextPage
+                    hasPreviousPage
+                    startCursor
+                    endCursor
                 }
             }
+        }
+    }
 
-            fragment RepoPermissionsInfoFields on PermissionsInfo {
-                syncedAt
-                updatedAt
-                unrestricted
-            }
-        `,
-        { repo }
-    ).pipe(
-        map(dataOrThrowErrors),
-        map(data => {
-            if (!data.node) {
-                throw new Error('Repository not found')
-            }
-            if (data.node.__typename !== 'Repository') {
-                throw new Error(`Node is a ${data.node.__typename}, not a Repository`)
-            }
-            return data.node.permissionsInfo
-        })
-    )
-}
+    fragment PermissionsInfoUserFields on PermissionsInfoUserNode {
+        id
+        reason
+        updatedAt
+        user {
+            id
+            username
+            displayName
+            email
+            avatarURL
+        }
+    }
+`