+++ internal/search/job/jobutil/enterprise.go
@@ -13,8 +13,9 @@ import (
 )
 
 type EnterpriseJobs interface {
-	FileHasOwnerJob(child job.Job, includeOwners, excludeOwners []string) job.Job
-	SelectFileOwnerJob(child job.Job) job.Job
+	FileHasOwnerJob(child job.Job, features *search.Features, includeOwners, excludeOwners []string) job.Job
+	SelectFileOwnerJob(child job.Job, features *search.Features) job.Job
+
 	SymbolRelationshipSearchJob(relationship query.SymbolRelationship, rawSymbolSearch job.Job) job.Job
 }
 
@@ -24,12 +25,12 @@ func NewUnimplementedEnterpriseJobs() EnterpriseJobs {
 
 type enterpriseJobs struct{}
 
-func (e *enterpriseJobs) FileHasOwnerJob(child job.Job, includeOwners, excludeOwners []string) job.Job {
-	return NewUnimplementedJob("file:has.owner searches are not available on this instance")
+func (e *enterpriseJobs) FileHasOwnerJob(child job.Job, features *search.Features, includeOwners, excludeOwners []string) job.Job {
+	return NewUnimplementedJob("`file:has.owner` searches are not available on this instance")
 }
 
-func (e *enterpriseJobs) SelectFileOwnerJob(child job.Job) job.Job {
-	return NewUnimplementedJob("file:select.owners searches are not available on this instance")
+func (e *enterpriseJobs) SelectFileOwnerJob(child job.Job, features *search.Features) job.Job {
+	return NewUnimplementedJob("`select:file.owners` searches are not available on this instance")
 }
 
 func (e *enterpriseJobs) SymbolRelationshipSearchJob(relationship query.SymbolRelationship, rawSymbolSearch job.Job) job.Job {