+++ internal/search/types.go
@@ -388,6 +388,7 @@ type RepoOptions struct {
 	UseIndex       query.YesNoOnly
 	HasFileContent []query.RepoHasFileContentArgs
 	HasKVPs        []query.RepoKVPFilter
+	HasTopics      []query.RepoHasTopicPredicate
 
 	// ForkSet indicates whether `fork:` was set explicitly in the query,
 	// or whether the values were set from defaults.
@@ -471,6 +472,18 @@ func (op *RepoOptions) Tags() []otlog.Field {
 			add(trace.Scoped(fmt.Sprintf("hasKVPs[%d]", i), nondefault...))
 		}
 	}
+	if len(op.HasTopics) > 0 {
+		for i, arg := range op.HasTopics {
+			nondefault := []otlog.Field{}
+			if arg.Topic != "" {
+				nondefault = append(nondefault, otlog.String("topic", arg.Topic))
+			}
+			if arg.Negated {
+				nondefault = append(nondefault, otlog.Bool("negated", arg.Negated))
+			}
+			add(trace.Scoped(fmt.Sprintf("hasTopics[%d]", i), nondefault...))
+		}
+	}
 	if op.ForkSet {
 		add(otlog.Bool("forkSet", op.ForkSet))
 	}
@@ -544,6 +557,16 @@ func (op *RepoOptions) String() string {
 			}
 			if arg.Negated {
 				fmt.Fprintf(&b, "HasKVPs[%d].negated: %t\n", i, arg.Negated)
+			}
+		}
+	}
+	if len(op.HasTopics) > 0 {
+		for i, arg := range op.HasTopics {
+			if arg.Topic != "" {
+				fmt.Fprintf(&b, "HasTopics[%d].topic: %s\n", i, arg.Topic)
+			}
+			if arg.Negated {
+				fmt.Fprintf(&b, "HasTopics[%d].negated: %t\n", i, arg.Negated)
 			}
 		}
 	}